Place cursor correctly before result overlay
When point is right before the result overlay, the cursor is displayed
after the overlay. Setting this text property fixes that, placing the
cursor in the place that matches the actual point position.

There might be other overlays that should get a similar treatment. I'll
leave that to you to figure out.